Combined pitch_tracker and backend interfaces
