+++ backend/SConscript
@@ -4,5 +4,5 @@ Import( 'env' )
 
 env = env.Clone()
 env.MergeFlags( "-I# -I#/libcore -I#/backend" )
-env.StaticLibrary( target="jackmix_backend", source=["jack_backend.cpp", "peak_tracker.cpp"] )
+env.StaticLibrary( target="jackmix_backend", source=["jack_backend.cpp"] )
